# src/main/java/com/teligen/acg/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "微服务工程代码生成器,生成Dao、MapperXml、Service、Controller、Vo、Feign、Swagger的JSON脚本",
    "version": "1.0.0",
    "title": "微服务工程代码生成器",
    "termsOfService": "https://github.com/guobk/",
    "contact": {
      "email": "836571084@qq.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
  "tags": [
        {
      "name": "UserController",
      "description": "UserController"
    },
    {
      "name": "UserController",
      "description": "UserController"
    },
    {
      "name": "UserController",
      "description": "UserController"
    },
    {
      "name": "UserController",
      "description": "UserController"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/user": {
      "get": {
        "tags": [
          "UserController"
        ],
        "summary": "查询所有User",
        "description": "查询所有User方法详情",
        "operationId": "findAllUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "查询所有User"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«User»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "UserController"
        ],
        "summary": "添加User",
        "description": "添加User方法详情",
        "operationId": "addUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "添加User方法详情",
            "required": true
              ,
              "schema": {
                "$ref": "#/definitions/User"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "添加User"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      }
    },
    "/user/{id}": {
      "delete": {
        "tags": [
          "UserController"
        ],
        "summary": "根据ID删除User",
        "description": "根据ID删除User方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除User方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除User"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "UserController"
        ],
        "summary": "根据ID修改User",
        "description": "根据ID修改User方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改User方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "User",
            "description": "传入User的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/User"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改User"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "UserController"
        ],
        "summary": "根据ID查询User",
        "description": "根据ID查询User方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改User方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改User"
            ,
            "schema": {
              "$ref": "#/definitions/Result«User»"
            }
          }
        }
      }
    },
    "/user/search": {
      "post": {
        "tags": [
          "UserController"
        ],
        "summary": "不带分页条件搜索User",
        "description": "不带分页条件搜索User方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "User",
            "description": "传入User的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/User"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索User"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«User»»"
            }
          }
        }
      }
    },
    "/user/search/{page}/{size}": {
      "get": {
        "tags": [
          "UserController"
        ],
        "summary": "分页列表查询User",
        "description": "分页列表查询User方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索User"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«User»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "UserController"
        ],
        "summary": "分页条件搜索User",
        "description": "分页条件搜索User方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "User",
            "description": "传入User的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/User"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索User"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«User»»"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Result": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "object",
          "description": "逻辑数据"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
    "Result«List«User»»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "array",
          "description": "逻辑数据",
          "items": {
            "$ref": "#/definitions/User"
          }
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
    "Result«User»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "description": "逻辑数据",
          "$ref": "#/definitions/User"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "description": ""
        },
        "userid": {
          "type": "string",
          "description": ""
        },
        "username": {
          "type": "string",
          "description": ""
        },
        "password": {
          "type": "string",
          "description": ""
        }
      },
      "description": "User"
    }
  },
  "externalDocs": {

  }
}